List for adding new music (#44)
This document contains a list as well as some general descriptions of new music this game could need (if someone is willing to work on it) and give direction on how to approach new songs to ensure quiality, consitency and theming! Further updates if need be shall be made at a later date!