More improvements to Tileset.md (#46)
As I've been creating and experimenting with tilesets for a long while now, I've decided to contribute to the tileset.md page, adding a more detailed explanation on how tile attributes work, which attributes are available and why some attributes are the number they have. I also added a guide on how to create special object tiles like coins, bonus blocks and bricks, as the attributes that made these behavior no longer work.